backend/requirements.txt
python-telegram-bot==20.7
flask==2.3.0
gunicorn==21.2.0